kafka-consumer-service/requirements.txt
kafka-python
cython
clickhouse-connect